Impl `openDocumentFileWithResult` and split src
