--- lib/media_store.dart
@@ -1,2 +1,0 @@
-export './src/media_store/media_store.dart';
-export './src/media_store/media_store_collection.dart';
