--- lib/saf.dart
@@ -1,5 +1,0 @@
-export './src/saf/document_bitmap.dart';
-export './src/saf/document_file.dart';
-export './src/saf/document_file_column.dart';
-export './src/saf/saf.dart';
-export './src/saf/uri_permission.dart';
