+++ lib/shared_storage.dart
@@ -1,5 +1,3 @@
 library shared_storage;
 
-export './environment.dart';
-export './media_store.dart';
-export './saf.dart';
+export 'src/barrel.dart';
