+++ lib/src/environment/barrel.dart
@@ -1,0 +1,2 @@
+export './environment.dart';
+export './environment_directory.dart';
