+++ lib/src/media_store/barrel.dart
@@ -1,0 +1,2 @@
+export './media_store.dart';
+export './media_store_collection.dart';
