+++ lib/src/saf/models/document_bitmap.dart
