+++ lib/src/saf/models/uri_permission.dart
